requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pydantic
